fix(server): replace `gameId` with `gameID`
This is required to make the tests pass. I think `gameId` is still the
correct naming choice, however.

This patch also includes a fix to return `null` if a game does not exist
in the data store instead of throwing an exception. When a user tries to
update a game that does not exist, we throw an exception. When a user
tries to get a game that does not exist, we simply return `null`.

Ref: https://cs240winter2024.slack.com/archives/C06CJDSTNH1/p1709170782556829