refactor(server): move endpoint logic to services
This patch moves the core endpoint logic to service classes that throw
`ExceptionWithStatusCode` to indicate what status code should be
forwarded via the REST API to the client.

In my opinion, this is somewhat useless abstraction, but the class wants
me to do this and add unit tests for the service methods, so I will.

Ref: https://github.com/softwareconstruction240/softwareconstruction/blob/main/chess/3-web-api/web-api.md#service-classes